🚀 feat(user-data.ts): add allowedTenants property to UserData interface 🔥 chore(home.component.html): comment out app-open-ai component 🚀 feat(smz-ui-config-builder.ts): add allowTenantSwitching method to SmzUiAuthenticationLoginBuilder 🔖 chore(ngx-smz-ui): update version to 15.1.0 🚀 feat(authentication-login-builder.ts): add mockedUserId property to SmzUiAuthenticationLoginBuilder 🚀 feat(ui-builder.ts): add allowTenantSwitching property to SmzUiBuilder 🐛 fix(pretty-json.pipe.ts): shorten long strings to 150 characters 🐛 fix(auth.interceptor.ts): use AUTHORIZATION_HEADER instead of AUTHENTICATION_HEADER 🚀 feat(auth.service.ts): add mockedUserId property to login request The changes add a new allowedTenants property to the UserData interface, which will be used to store the list of
🚀 feat(api): add support for mocked user id header
🚀 feat(api): add support for switching tenants
🚀 feat(authentication.service.ts): add switchTenant method
🚀 feat(tenants.selectors.ts): add userAllowedTenants and currentTenant selectors
The mocked user id header is now supported, which allows for testing of user-specific functionality. The switchTenant method has been added to the AuthenticationService, which allows users to switch between tenants. The userAllowedTenants and currentTenant selectors have been added to the TenantsSelectors, which allows for easier access to the allowed tenants and the current tenant.

🚀 feat(tenant-switch): add tenant switch component to topbar
The SmzTenantSwitchComponent has been added to the topbar of both the Athena and Hephaestus layouts. This allows users to switch between tenants if the allowTenantSwitching configuration is set to true.

🚀 feat(topbar-actions): add SmzTenantSwitchComponent to topbar-actions module
🚀 feat(login): add support for tenant switch and development credentials
🚀 feat(authentication.actions): add SwitchTenant action
The SmzTenantSwitchComponent is now added to the topbar-actions module, which allows the user to switch between tenants. The login component now supports tenant switching and development credentials. The SwitchTenant action is added to the authentication.actions file to handle tenant switching.

🚀 feat(authentication.state.ts): add SwitchTenant action
The `SwitchTenant` action is added to the `AuthenticationState` class. This action handles the switching of tenants by calling the `switchTenant` method of the `AuthenticationService`. Upon success, the `RemoteLoginSuccess` action is dispatched with the new access and refresh tokens. The page is then reloaded to reflect the changes.